api/requirements.txt
fastapi==0.100.0
Pillow==9.5.0
pydantic==1.10.9
torch==2.0.1
torchvision==0.15.2
uvicorn==0.22.0
